anexi-backend/requirements.txt
fastapi
uvicorn
SQLAlchemy>=1.4
psycopg2-binary
pydantic
pydantic[email]
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt<5
python-multipart
python-dotenv
celery>=5.3
redis>=5.0
httpx>=0.27
prometheus-client>=0.20
opentelemetry-api>=1.25
opentelemetry-sdk>=1.25
opentelemetry-exporter-otlp-proto-http>=1.25
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-httpx>=0.46b0
opentelemetry-instrumentation-redis>=0.46b0
opentelemetry-instrumentation-celery>=0.46b0
